// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const GoalType = {
  LONG_TERM: 'LONG_TERM',
  QUARTERLY: 'QUARTERLY',
  MONTHLY: 'MONTHLY',
  WEEKLY: 'WEEKLY',
  DAILY: 'DAILY'
} as const

export type GoalType = (typeof GoalType)[keyof typeof GoalType]


export const Priority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const Status = {
  NOT_STARTED: 'NOT_STARTED',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
  BLOCKED: 'BLOCKED'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const Platform = {
  LEETCODE: 'LEETCODE',
  CODEFORCES: 'CODEFORCES',
  GFG: 'GFG',
  OTHER: 'OTHER'
} as const

export type Platform = (typeof Platform)[keyof typeof Platform]


export const Difficulty = {
  EASY: 'EASY',
  MEDIUM: 'MEDIUM',
  HARD: 'HARD'
} as const

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty]


export const ProjectKind = {
  COMPANY_WEBSITE: 'COMPANY_WEBSITE',
  FLAGSHIP: 'FLAGSHIP',
  OTHER: 'OTHER'
} as const

export type ProjectKind = (typeof ProjectKind)[keyof typeof ProjectKind]


export const RiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH'
} as const

export type RiskLevel = (typeof RiskLevel)[keyof typeof RiskLevel]


export const AppStage = {
  APPLIED: 'APPLIED',
  OA: 'OA',
  INTERVIEW: 'INTERVIEW',
  OFFER: 'OFFER',
  REJECTED: 'REJECTED'
} as const

export type AppStage = (typeof AppStage)[keyof typeof AppStage]


export const LearningType = {
  BOOK: 'BOOK',
  COURSE: 'COURSE',
  ARTICLE: 'ARTICLE',
  VIDEO: 'VIDEO'
} as const

export type LearningType = (typeof LearningType)[keyof typeof LearningType]